+++ workspaces/announcements/plugins/announcements/src/components/AnnouncementsCard/AnnouncementsCard.tsx
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import React from 'react';
 import { InfoCard, InfoCardVariants } from '@backstage/core-components';
 import { useRouteRef } from '@backstage/core-plugin-api';
 import { useAnnouncementsTranslation } from '@backstage-community/plugin-announcements-react';