+++ workspaces/rbac/plugins/rbac-backend/src/service/enforcer-delegate.ts
@@ -162,16 +162,16 @@ export class EnforcerDelegate implements RoleEventEmitter<RoleEvents> {
   ): Promise<string[][]> {
     const tempModel = newModelFromString(MODEL);
 
-    const filterArgs: Record<string, string>[] = [];
     const filterObj: Record<string, string> = { ptype: 'p' };
     for (let i = 0; i < filter.length; i++) {
-      filterObj[`v${i + fieldIndex}`] = filter[i];
-      filterArgs.push(filterObj);
+      if (filter[i]) {
+        filterObj[`v${i + fieldIndex}`] = filter[i];
+      }
     }
 
     await (this.enforcer.getAdapter() as FilteredAdapter).loadFilteredPolicy(
       tempModel,
-      filterArgs,
+      [filterObj],
     );
 
     return await tempModel.getPolicy('p', 'p');
@@ -183,16 +183,16 @@ export class EnforcerDelegate implements RoleEventEmitter<RoleEvents> {
   ): Promise<string[][]> {
     const tempModel = newModelFromString(MODEL);
 
-    const filterArgs: Record<string, string>[] = [];
     const filterObj: Record<string, string> = { ptype: 'g' };
     for (let i = 0; i < filter.length; i++) {
-      filterObj[`v${i + fieldIndex}`] = filter[i];
-      filterArgs.push(filterObj);
+      if (filter[i]) {
+        filterObj[`v${i + fieldIndex}`] = filter[i];
+      }
     }
 
     await (this.enforcer.getAdapter() as FilteredAdapter).loadFilteredPolicy(
       tempModel,
-      filterArgs,
+      [filterObj],
     );
 
     return await tempModel.getPolicy('g', 'g');