+++ workspaces/redhat-argocd/plugins/argocd/src/components/AppStatus/AppHealthStatus.tsx
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import React from 'react';
+import type { FC } from 'react';
 
 import { Chip } from '@material-ui/core';
 
@@ -23,7 +23,7 @@ import {
 } from '@backstage-community/plugin-redhat-argocd-common';
 import { AppHealthIcon } from './StatusIcons';
 
-const AppHealthStatus: React.FC<{ app: Application; isChip?: boolean }> = ({
+const AppHealthStatus: FC<{ app: Application; isChip?: boolean }> = ({
   app,
   isChip = false,
 }) => {