+++ workspaces/3scale/plugins/3scale-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## @janus-idp/backstage-plugin-3scale-backend [1.8.0](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-3scale-backend@1.7.1...@janus-idp/backstage-plugin-3scale-backend@1.8.0) (2024-07-25)
 
+## 3.4.0
+
+### Minor Changes
+
+- 5447deb: Backstage version bump to v1.38.1
+
 ## 3.3.2
 
 ### Patch Changes