+++ workspaces/announcements/plugins/announcements-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-announcements-common
 
+## 0.3.2
+
+### Patch Changes
+
+- 11be6bb: chore(deps): Upgrade to Backstage 1.38
+
 ## 0.3.1
 
 ### Patch Changes